ログアウトとデータ削除の時確認ダイアログが表示

+++ mytodoapp/lib/widgets/post_item.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
@@ -28,7 +29,7 @@ class PostItem extends StatelessWidget {
                 (data['done'] ?? false) ? TextDecoration.lineThrough : null,
           ),
         ),
-        subtitle: Text(data['email']),
+        // subtitle: Text(data['email']),
         leading: Checkbox(
           value: data['done'] ?? false,
           onChanged: (val) {
@@ -55,17 +56,51 @@ class PostItem extends StatelessWidget {
                   ),
                   IconButton(
                     icon: const Icon(Icons.delete),
-                    onPressed: () async {
-                      await FirebaseFirestore.instance
-                          .collection('posts')
-                          .doc(document.id)
-                          .delete();
+                    onPressed: () {
+                      showDeleteConfirmDialog(context);
                     },
                   ),
                 ],
               )
             : null,
       ),
+    );
+  }
+
+  void showDeleteConfirmDialog(BuildContext context) {
+    showDialog(
+      context: context,
+      builder: (context) {
+        return CupertinoAlertDialog(
+          title: const Text('削除確認'),
+          content: const Text('この投稿を本当に削除しますか？'),
+          actions: [
+            CupertinoDialogAction(
+              child: const Text(
+                'キャンセル',
+                style: TextStyle(color: Colors.blueAccent),
+              ),
+              onPressed: () {
+                Navigator.pop(context);
+              },
+            ),
+            CupertinoDialogAction(
+              child: const Text(
+                '削除',
+                style: TextStyle(color: Colors.red),
+              ),
+              isDestructiveAction: true,
+              onPressed: () async {
+                Navigator.pop(context); // ダイアログを閉じる
+                await FirebaseFirestore.instance
+                    .collection('posts')
+                    .doc(document.id)
+                    .delete();
+              },
+            ),
+          ],
+        );
+      },
     );
   }
 }
